Added X10 Command Line interface
An X10 command line interface has been added for use on a host Linux system
to interface to the X10 Master via i2c.  Changes were made to the Makefile
to allow for building the CLI.  Note that the compiler must be specified
when invoking the Makefile to build the CLI tool, for example:

`make CLI_CC=avr32-linux-gcc x10cli`

Additional CFLAGS and LDFLAGS can be specified via the CLI_CFLAGS and the
CLI_LDFLAGS variables (although these are optional).